+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Changes for developers
 * Changes to the service
   * Fetching class information will now also return information about implemented interfaces and used traits.
+  * Fetching class information will now also return information about direct children, direct implementors (if it's an interface) and direct users (if it's a trait).
 
 ## 0.5.4
 ### Features and enhancements